+++ docs/CHANGELOG.rst
@@ -6,8 +6,24 @@ All notable changes to the Vector SIL Kit project shall be documented in this fi
 
 The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <http://keepachangelog.com/en/1.0.0/>`_.
 
-[unreleased]
+
+[4.0.53] - UNRELEASED
 ---------------------
+
+Changed
+~~~~~~~
+
+- Harmonized user interface of commandline utilities Registry, Monitor, and SystemController.
+
+  All utilities listen to OS signals to detect termination request.
+
+  - Old: Monitor and Registry wait for the user to press `[Enter]`` to shutdown.
+  - New: Monitor and Registry can be requested to shutdown via ``[Ctrl]-[C]``. Registry option '-s, --use-signal-handler' is deprecated.
+
+  All utilities no longer require user interaction.
+
+  - Old: Registry, Monitor and System Controller prompt the user by default to press `[Enter]`` to end process after shutdown.
+  - New: All utilities terminate without prompting for user input. SystemController option '-ni, --non-interactive' is deprecated.
 
 
 [4.0.52] - 2024-09-02